requirements.txt
openai-whisper
faster-whisper
sounddevice
numpy
pyperclip
pyautogui
pywin32
keyboard
pystray
Pillow
whisperx
pydub
